#12 - DMS 연산을 위한 타입 변경

+++ object/src/main/java/com/_P/eureka/client/object/hub/dto/HubCreateDto.java
@@ -8,9 +8,9 @@ public class HubCreateDto {
 
     private String address;
 
-    private int latitude;
+    private Long latitude;
 
-    private int longitude;
+    private Long longitude;
 
     public Hub toEntity(){
         return Hub.builder()
